# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "MegaHack4",
    "description": "",
    "contact": {}
  },
  "host": "megahack5.brunoeleodoro.com",
  "basePath": "/api",
  "securityDefinitions": {},
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/auth/eletrodomesticos": {
      "get": {
        "description": "",
        "summary": "Get Eletrodomesticos",
        "tags": [
          "user"
        ],
        "operationId": "GetEletrodomesticos",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object"
            },
            "headers": {}
          }
        }
      },
      "post": {
        "description": "",
        "summary": "Atualizar Eletrodomesticos",
        "tags": [
          "user"
        ],
        "operationId": "AtualizarEletrodomesticos",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/AtualizarEletrodomesticosRequest"
            }
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object"
            },
            "headers": {}
          }
        }
      }
    },
    "/auth/leituras": {
      "get": {
        "description": "",
        "summary": "Get Leituras",
        "tags": [
          "user"
        ],
        "operationId": "GetLeituras",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object"
            },
            "headers": {}
          }
        }
      },
      "post": {
        "description": "",
        "summary": "Atualizar Leituras",
        "tags": [
          "user"
        ],
        "operationId": "AtualizarLeituras",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/AtualizarLeiturasRequest"
            }
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object"
            },
            "headers": {}
          }
        }
      }
    },
    "/auth/historico": {
      "get": {
        "description": "",
        "summary": "Get Historico",
        "tags": [
          "user"
        ],
        "operationId": "GetHistorico",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object"
            },
            "headers": {}
          }
        }
      },
      "post": {
        "description": "",
        "summary": "Atualizar Historico",
        "tags": [
          "user"
        ],
        "operationId": "AtualizarHistorico",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/AtualizarHistoricoRequest"
            }
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object"
            },
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
    "AtualizarEletrodomesticosRequest": {
      "title": "AtualizarEletrodomesticosRequest",
      "example": {
        "eletrodomesticos": [
          {
            "title": "Plasma 42",
            "horas": "5",
            "quantidade": "1"
          }
        ]
      },
      "type": "object",
      "properties": {
        "eletrodomesticos": {
          "type": "array",
          "items": {
            "type": "object"
          }
        }
      },
      "required": [
        "eletrodomesticos"
      ]
    },
    "AtualizarLeiturasRequest": {
      "title": "AtualizarLeiturasRequest",
      "example": {
        "leituras": [
          {
            "title": "Plasma 42",
            "horas": "5",
            "quantidade": "1"
          }
        ]
      },
      "type": "object",
      "properties": {
        "leituras": {
          "type": "array",
          "items": {
            "type": "object"
          }
        }
      },
      "required": [
        "leituras"
      ]
    },
    "AtualizarHistoricoRequest": {
      "title": "AtualizarHistoricoRequest",
      "example": {
        "historico": [
          {
            "title": "Plasma 42",
            "horas": "5",
            "quantidade": "1"
          }
        ]
      },
      "type": "object",
      "properties": {
        "historico": {
          "type": "array",
          "items": {
            "type": "object"
          }
        }
      },
      "required": [
        "historico"
      ]
    },
...
  },
...
}